Improves DH model
